# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "3CM API Blog",
        "description": "API documentation for the 3CM Blog project",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/newsletter/subscribers": {
            "get": {
                "tags": [
                    "Admin Newsletter"
                ],
                "summary": "List all newsletter subscribers (Admin)",
                "description": "Display a listing of subscribers (Admin).",
                "operationId": "450138c2ae4e3a0f8e336a796c089ed2",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/NewsletterSubscriber"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/newsletter/subscribe": {
            "post": {
                "tags": [
                    "Newsletter"
                ],
                "summary": "Subscribe to the newsletter",
                "description": "Handle a new subscription.",
                "operationId": "0f0d2c985d1bb0830af23d94fb2ec713",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscribed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/newsletter/unsubscribe": {
            "post": {
                "tags": [
                    "Newsletter"
                ],
                "summary": "Unsubscribe from the newsletter",
                "description": "Handle unsubscription.",
                "operationId": "18f53645289ab12902a0570a4f93ff08",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "random_token_string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Unsubscribed successfully"
                    },
                    "404": {
                        "description": "Subscriber not found or invalid token"
                    }
                }
            }
        },
...
        "/api/v1/admin/settings": {
            "put": {
                "tags": [
                    "Admin Settings"
                ],
                "summary": "Update site settings (Admin)",
                "description": "Update settings (Admin).",
                "operationId": "a8bc1e9068d86b2f2228744bf7d178ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "site_name": "Mamilo Blog",
                                "site_description": "A modern blog platform"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Settings updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NewsletterSubscriber": {
                "required": [
                    "email",
                    "unsubscribe_token"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "subscribed_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}